Add text.
Started writing comments and observations about the first part of the
experiments (degree distr. parameters and log-likelihood estimation).
Also added some tags to the chunks of code to hide them (in some cases)
to avoid showing the code and just making the text visible as well as
the tables generated. We should check how to organize the structure.